MDL-18375 calendar: removed code from the base class and only introduced functionality for the date form elements
In the original patch we are overwriting core functions responsible for saving time in the
database in the calendar base class. This is very dangerous, we do not want to touch these
functions as it could cause a complete fubar of the database. There are places we are forcing
the use of the gregorian calendar as we are passing dates generated by the PHP date function,
where as sometimes we pass dates from usergetdate (which was being overwritten to return the
date specific to the calendar). We can not expect third party modules to change the calendar
system depending on the format they pass to these functions. In future commits I will be
expanding the base class when needed and removing any changes to the core functions when
related to the commit.