MDL-18375 calendar: prevent usage of define_after_data when creating a date/time profile field
If we are adding a new profile field then the years have already been set by setDefault
to the correct years in the used calendar type. We only want to execute the code in
define_after_data when we have the years in the DB, which are saved in Gregorian,
that need converting to the equivalent for the currently used calendar type.