MDL-18375 calendar: fixed the issue where changing calendar types and resaving the year in the date/time field caused the year to decrease.
For example, a user may create a date/time profile field and set the 'Start year' field to
'2012'. Another user, using the Hijri calendar, may edit the name of this field. They visit
the settings page and in this case the date '1/1/2012' is converted to '7/2/1433' in Hijri.
So, the year '1433' is then displayed. The user then changes the name of this field and
saves the form. This is where the issue occurs, as the date '1/1/1433' is converted into
Gregorian, which converts to the year '26/09/2011', so the year '2011' is saved in the DB,
not '2012'.