MDL-36941 message_popup: do not remove message_popup table
Partial revert "MDL-36941 message_popup: removed 'message_popup' table"

This is a partial revert of commit 123631ad4718141627d77027dbef0448bae22ac8.

Most of the logic was reverted but there were a few adjustments due
to other changes by MDL-36941.

Note - this fix was done on integration.git and as a result no
upgrade step was added to re-introduce the table, sorry! This
means for those on integration.git you will have to re-install
your Moodle site so the 'message_popup' table is present.