MDL-36941 message_popup: add table to store notifications
We are migrating data from the 'message_popup' table to the
new table 'message_popup_notifications'.

We no longer need an event observer. The observer was used
because the old system would move a message from the 'message'
table to the 'message_read' table when it was viewed and the
'messageid' field needed to be updated. Now, we only have one
table 'notifications' so there is no need for this.