MDL-27520 core: allow retrieval of files from grade files areas
Also added 'history' filearea for the history report where files
will be copied to if the site has enabled this feature.

We will also be storing files in the module context, not course,
so that check was removed. The reason for this decision is because
a user may backup a course but choose not to include a certain
activity. In that case the files related to that activity should not
be chosen.